Fixed a warning when compiling for iOS devices

+++ JPImgurKit/Models/JPImgurAccount.m
@@ -48,7 +48,7 @@
 {
     return [NSString stringWithFormat:
             @"accountID: %lu; url: \"%@\"; bio: \"%@\"; reputation: %.2f; created: %@",
-            _accountID, _url, _bio, _reputation, _created];
+            (unsigned long)_accountID, _url, _bio, _reputation, _created];
 }
 
 @end
